requirements.txt
numpy>=1.22.0,<2.0.0
pandas>=1.4.0
scikit-learn>=1.0.0
websockets>=10.0
torch>=1.9.0
librosa>=0.8.1
transformers>=4.11.0
soundfile>=0.10.3
ebooklib>=0.17.1
python-Levenshtein
googletrans==4.0.0-rc1
pillow>=10.0.0